Add initial support for multiple apis in client module
A new enum contains all supported APIs. Currently only lichess.org and chess.com. The enum builds requests for the supported endpoints, returns a result with an error if the endpoint is not supported. Chess.com is the hardcoded default API for now as we are still missing a API/host/website/other CLI parameter and specific finding operations for other APIs besides chess.com.