Implement DisplayableChessGame trait to support multiple Game structs
Since each API (chess.com, lichess.org, and others) return different fields when requesting a game, we need a common interface for all of them. That's where the DisplayableChessGame trait comes in: it ensures any struct implements the ChessGame, Serializable, Debug, and Clone traits to ensure they can be used to print out PGN and JSON. Some more work is required to support lichess.org though, as the finder is still not ready to process anything but chess.com responses.